Bug: User created in sign-up form but full name remains blank
auth.createUserWithEmailAndPassword returns promise, so we need to await for this function to return but in sign-up.component.jsx we din't await and immediately after calling this function we called creaeUserProfileDocument in which user object remains null and user not creaed. and later when it created by App.js onAuthStateChaged call back there is no displayName property available so it creates user with blank displayName.